move rendering into app to keep react state, distinguish between old and new render path
