+++ src/components/pages/search.tsx
@@ -1,10 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { CookieBar } from '../content/cookie-bar'
 import { HeadTags } from '../head-tags'
-import { Footer } from '../navigation/footer'
-import { Header } from '../navigation/header'
 import { SearchResults } from '@/components/navigation/search-results'
 import { useInstanceData } from '@/contexts/instance-context'
 
@@ -12,13 +9,10 @@ export function Search() {
   const { strings } = useInstanceData()
   return (
     <>
-      <Header onSearchPage />
       <HeadTags data={{ title: `Serlo.org - ${strings.header.search}` }} />
       <StyledSearchResults>
         <div className="gcse-searchresults"></div>
       </StyledSearchResults>
-      <Footer />
-      <CookieBar />
     </>
   )
 }
